feat: split component and style && add themaProvider
