+++ src/components/ChooseFriend/Choose_Friend.tsx
@@ -1,0 +1,10 @@
+import { Message, PlusButton, Wrapper } from '@/components/ChooseFriend/ChooseFriend.styes.ts';
+
+export default function Choose_Friend() {
+  return (
+    <Wrapper>
+      <PlusButton>+</PlusButton>
+      <Message>선물한 친구를 선택해 주세요.</Message>
+    </Wrapper>
+  );
+}
